+++ android/domain/src/main/java/com/bonobono/domain/usecase/community/DeleteArticleUseCase.kt
@@ -6,7 +6,7 @@ import javax.inject.Inject
 class DeleteArticleUseCase @Inject constructor(
     private val communityRepository: CommunityRepository
 ) {
-    suspend operator fun invoke(type: String, articleId: Int) {
+    suspend operator fun invoke(type: String, articleId: Long) {
         communityRepository.deleteArticle(type, articleId)
     }
 }
