+++ android/domain/src/main/java/com/bonobono/domain/usecase/community/UpdateArticleLikeUseCase.kt
@@ -7,7 +7,7 @@ import javax.inject.Inject
 class UpdateArticleLikeUseCase @Inject constructor(
     private val communityRepository: CommunityRepository
 ){
-    suspend operator fun invoke(type: String, articleId: Int): NetworkResult<Unit> {
+    suspend operator fun invoke(type: String, articleId: Long): NetworkResult<Unit> {
         return communityRepository.updateArticleLike(type, articleId)
     }
 }
